Add an iterator over the segment groups
Dealing with the index based group ranges is not ergonomic. Therefore
this commits adds an iterator that returns all the groups as views of
segment slices. Also segments that are not part of any groups become
pseudo-groups of single segments. The splits component should be able to
build on top of this.